restore & move to mask
